reorganization of code base
