--- output.xml
@@ -1,1 +1,0 @@
-<XModelerPackage path="Root::Modell" version="4"><Imports /><Model name="Root::Modell"><addMetaClass abstract="false" level="1" maxLevel="1" name="person" package="Root::Modell" singleton="false" /><addMetaClass abstract="false" level="1" maxLevel="1" name="Student" package="Root::Modell" singleton="false" /><addAttribute level="0" multiplicity="Seq{1,1,true,false}" name="firstName" package="Root::Modell" type="Root::XCore::String" class="Root::Modell::Student" /><addAttribute level="0" multiplicity="Seq{1,1,true,false}" name="lastName" package="Root::Modell" type="Root::XCore::String" class="Root::Modell::Student" /><addAttribute level="0" multiplicity="Seq{1,1,true,false}" name="matriculationNr" package="Root::Modell" type="Root::XCore::Integer" class="Root::Modell::Student" /><addAttribute level="0" multiplicity="Seq{1,1,true,false}" name="privateMailAddress" package="Root::Modell" type="Root::XCore::String" class="Root::Modell::Student" /><addAttribute level="0" multiplicity="Seq{1,1,true,false}" name="subjectTerm" package="Root::Modell" type="Root::XCore::Integer" class="Root::Modell::Student" /><addAttribute level="0" multiplicity="Seq{1,1,true,false}" name="universityMailAddress" package="Root::Modell" type="Root::XCore::String" class="Root::Modell::Student" /><addAttribute level="0" multiplicity="Seq{1,1,true,false}" name="universityTerm" package="Root::Modell" type="Root::XCore::Integer" class="Root::Modell::Student" /><changeParent new="Root::Modell::person" old="" package="Root::Modell" class="Root::Modell::Student" /><addMetaClass abstract="false" level="1" maxLevel="1" name="Employee" package="Root::Modell" singleton="false" /><addAttribute level="0" multiplicity="Seq{1,1,true,false}" name="function" package="Root::Modell" type="Root::XCore::String" class="Root::Modell::Employee" /><addAttribute level="0" multiplicity="Seq{1,1,true,false}" name="givenName" package="Root::Modell" type="Root::XCore::String" class="Root::Modell::Employee" /><addAttribute level="0" multiplicity="Seq{1,1,true,false}" name="mailaddress" package="Root::Modell" type="Root::XCore::String" class="Root::Modell::Employee" /><addAttribute level="0" multiplicity="Seq{1,1,true,false}" name="personalConsultationDay" package="Root::Modell" type="Root::XCore::String" class="Root::Modell::Employee" /><addAttribute level="0" multiplicity="Seq{1,1,true,false}" name="personalConsultationHours" package="Root::Modell" type="Root::XCore::String" class="Root::Modell::Employee" /><addAttribute level="0" multiplicity="Seq{1,1,true,false}" name="phoneNr" package="Root::Modell" type="Root::XCore::String" class="Root::Modell::Employee" /><addAttribute level="0" multiplicity="Seq{1,1,true,false}" name="surname" package="Root::Modell" type="Root::XCore::String" class="Root::Modell::Employee" /><addAttribute level="0" multiplicity="Seq{1,1,true,false}" name="telephoneConsultationDay" package="Root::Modell" type="Root::XCore::String" class="Root::Modell::Employee" /><addAttribute level="0" multiplicity="Seq{1,1,true,false}" name="telephoneConsultationHours" package="Root::Modell" type="Root::XCore::String" class="Root::Modell::Employee" /><changeParent new="Root::Modell::person" old="" package="Root::Modell" class="Root::Modell::Employee" /></Model><Diagrams><Diagram name="Model"><Instances><Instance hidden="false" path="Root::Modell::person" xCoordinate="0" yCoordinate="0" /><Instance hidden="false" path="Root::Modell::Student" xCoordinate="0" yCoordinate="0" /><Instance hidden="false" path="Root::Modell::Employee" xCoordinate="0" yCoordinate="0" /></Instances><Edges /><DiagramDisplayProperties SHOWCONCRETESYNTAX="true" SHOWCONSTRAINTREPORTS="true" SHOWCONSTRAINTS="true" SHOWDERIVEDATTRIBUTES="true" SHOWDERIVEDOPERATIONS="true" SHOWGETTERSANDSETTERS="false" SHOWISSUETABLEVISIBLE="false" SHOWMETACLASSNAME="false" SHOWOPERATIONS="true" SHOWOPERATIONVALUES="true" SHOWSLOTS="true" /><View name="Main View" tx="0.0" ty="0.0" xx="1.0" /></Diagram></Diagrams></XModelerPackage>
