+++ src/_archive/lexica.py
