--- test.xml
@@ -1,1 +1,0 @@
-<XModelerPackage path="Root::Model" version="4"><Imports /><Model name="Root::Model"><addMetaClass abstract="false" level="1" maxLevel="1" name="Monograph" package="Root::Model::" singleton="false" /><addAttribute level="0" multiplicity="Seq{1,1,true,false}" name="lastUpdate" package="Root::Model::" type="Root::XCore::String" class="Root::Model::Monograph" /><addAttribute level="0" multiplicity="Seq{1,1,true,false}" name="alter" package="Root::Model::" type="Root::XCore::Integer" class="Root::Model::Monograph" /><addAttribute level="0" multiplicity="Seq{1,1,true,false}" name="dautm" package="Root::Model::" type="Root::Auxiliary::Date" class="Root::Model::Monograph" /><addInstance abstract="false" level="0" maxLevel="0" name="Mono1" of="Root::Model::Monograph" package="Root::Model::" singleton="false" /><changeSlotValue package="Root::Model::" slotName="lastUpdate" valueToBeParsed="[116, 101, 115, 116].asString()" class="Root::Model::Mono1" /><changeSlotValue package="Root::Model::" slotName="alter" valueToBeParsed="1" class="Root::Model::Mono1" /><changeSlotValue package="Root::Model::" slotName="dautm" valueToBeParsed="Auxiliary::Date::createDate(2020,5,17)" class="Root::Model::Mono1" /><addInstance abstract="false" level="0" maxLevel="0" name="Mono2" of="Root::Model::Monograph" package="Root::Model::" singleton="false" /><changeSlotValue package="Root::Model::" slotName="lastUpdate" valueToBeParsed="[112, 105, 101, 114, 114, 101, 73, 115, 116, 68, 111, 111, 102].asString()" class="Root::Model::Mono2" /><changeSlotValue package="Root::Model::" slotName="alter" valueToBeParsed="1" class="Root::Model::Mono2" /><changeSlotValue package="Root::Model::" slotName="dautm" valueToBeParsed="Auxiliary::Date::createDate(2020,5,17)" class="Root::Model::Mono2" /></Model><Diagrams><Diagram name="Model"><Instances><Instance hidden="false" path="Root::Model::::Monograph" xCoordinate="0" yCoordinate="0" /><Instance hidden="false" path="Root::Model::::Mono1" xCoordinate="0" yCoordinate="0" /><Instance hidden="false" path="Root::Model::::Mono2" xCoordinate="0" yCoordinate="0" /></Instances><Edges /><DiagramDisplayProperties SHOWCONCRETESYNTAX="true" SHOWCONSTRAINTREPORTS="true" SHOWCONSTRAINTS="true" SHOWDERIVEDATTRIBUTES="true" SHOWDERIVEDOPERATIONS="true" SHOWGETTERSANDSETTERS="false" SHOWISSUETABLEVISIBLE="false" SHOWMETACLASSNAME="false" SHOWOPERATIONS="true" SHOWOPERATIONVALUES="true" SHOWSLOTS="true" /><View name="Main View" tx="0.0" ty="0.0" xx="1.0" /></Diagram></Diagrams></XModelerPackage>
